LayoutTests/http/wpt/webauthn/public-key-credential-get-failure-hid-silent.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[get]] with malicious payload in a mock hid authenticator. 
PASS PublicKeyCredential's [[get]] with invalid credential in a mock hid authenticator. 
PASS PublicKeyCredential's [[get]] with authenticator downgrade in a mock hid authenticator. 
PASS PublicKeyCredential's [[get]] with authenticator downgrade in a mock hid authenticator. 2 
